change some int types to size_t.

+++ assets/max.c
@@ -1,4 +1,4 @@
-int max(const int *arr, int size) {
+int max(const int *arr, size_t size) {
     int temp = arr[0];
 
     for (int i = 0; i < size; i++)
