+++ services/queens/evalQueens.c
@@ -1,14 +1,14 @@
 #include "../../headers/mainHeaders.h"
 
-int *evalQueens(const int *matrix, int size);
+int *evalQueens(const int *matrix, size_t size);
 
-int evalSolQueen(const int *sol, int size);
+int evalSolQueen(const int *sol, size_t size);
 
 /*
  * To find diagonal collisions, just find differences between each queen's location (row, col) with other queens; If
  * differences are the same so these queens are met each other.
  * */
-int *evalQueens(const int *matrix, int size) {
+int *evalQueens(const int *matrix, size_t size) {
     int *tempArr = (int *) calloc(size, sizeof(int));
 
     for (int i = 0; i < size; i++)
@@ -18,7 +18,7 @@ int *evalQueens(const int *matrix, int size) {
 }
 
 //// O(n) new method
-int evalSolQueen(const int *sol, int size) {
+int evalSolQueen(const int *sol, size_t size) {
     int collision = 0, i = 0, j = i + 1;
 
     while (i < size - 2) {
@@ -37,7 +37,7 @@ int evalSolQueen(const int *sol, int size) {
 }
 
 //// O(n^2) old method
-//int evalSolQueen(const int *sol, int size) {
+//int evalSolQueen(const int *sol, size_t size) {
 //    int collision = 0;
 //
 //    for (int i = 0; i < size - 1; i++)
